Truncate long network names (#12058)
In some languages, network names are longer than the visible width of
the network dropdown. In these cases, the currently selected network will
overflow the bounds of the network dropdown container.

This commit addresses this issue so that the dropdown container will
grow and shrink appropriately, truncating the currently selected network
if needed.